LayoutTests/platform/wpe/storage/filesystemaccess/sync-access-handle-basics-worker-expected.txt

[Worker] This test checks basic funtionalities of FileSystemSyncAccessHandle.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/sync-access-handle-basics.js
PASS [Worker] fileHandle.name is "sync-access-handle.txt"
PASS [Worker] fileHandle.kind is "file"
FAIL [Worker] NotSupportedError: Backend does not support this operation
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
